% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_pitch_tracking.R
\name{read_data_pitch_tracking}
\alias{read_data_pitch_tracking}
\title{Read pitch tracking data}
\usage{
read_data_pitch_tracking()
}
\value{
a table of primary fastball velo indexed by pitcher_id and year
}
\description{
This function reads pitch tracking data from three different sources and consolidates it into
one row per pitcher-year.
}
