% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_ucl_injury.R
\name{read_data_ucl_injury}
\alias{read_data_ucl_injury}
\title{Read UCL injury data}
\usage{
read_data_ucl_injury()
}
\value{
a table of UCL injuries indexed by player_id and year
}
\description{
This function reads UCL injury data from multiple sources and combines them.
}
